api/requirements.txt
Django==5.1.5
djangorestframework-simplejwt==5.4.0
djangorestframework==3.15.2
django-cors-headers==4.7.0
django-filter==25.1
black==25.1.0
requests==2.32.3
pytest==8.4.1
coverage==7.9.2
pytest-cov
redis==5.0.1
django-redis==5.4.0